Safari-611.1.10/LayoutTests/inspector/page/overrideSetting-MockCaptureDevicesEnabled-expected.txt

Tests for the Page.overrideSetting command.


== Running test suite: Page.overrideSetting
-- Running test case: Page.overrideSetting.MockCaptureDevicesEnabled
Overriding MockCaptureDevicesEnabled to false...
Mock use check: PASS
Overriding MockCaptureDevicesEnabled to true...
Mock use check: PASS
Removing MockCaptureDevicesEnabled override...
Mock use check: PASS
